fix: some fixes

+++ src/scoreinterface.h
@@ -20,10 +20,10 @@ public:
 
   Q_INVOKABLE BestScoreModel *getBestScores() { return &m_bestScoreModel; }
   Q_INVOKABLE void resetBestScores();
-  Q_INVOKABLE void addBestScore(uint64_t score);
+  Q_INVOKABLE void addBestScore(int score);
 
 private:
-  std::vector<std::tuple<uint64_t, QDateTime>> m_bestScores;
+  std::vector<std::tuple<int, QDateTime>> m_bestScores;
   BestScoreModel m_bestScoreModel;
   fs::path m_saveFilename;
 };
